move pah and line tbl from cafe.py to cafe_io.py
